Move compare method into PartialOrd trait
What's funny is the way I structured the function is the exact same as
the trait, event to the point that the return type is an enum with
Equals, Greater, and Less. I made them capitals because I thought that was
the convention for Rust enums but I'll check again after seeing that.
Anyway because the structures were the same I could just copy the func-
tion over and just needing to change the return type. Yay